requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.6
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.14.0
idna==3.6
psycopg==3.1.12
psycopg-binary==3.1.12
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
